active learning 0.1 with image foveation and expandable class size
The Fovea class generates masks which can be applied to an image to create a foveated
view similar to the human eye.

The ActiveLearner class takes images (masked or unmasked) and trains on
images one by one and keeps a memory of some past frames to reduce the
risk of catastrophic forgetting. The ActiveLearner can update/grow the
final layer of the model while preserving their weights. This way, the
original predictive behaviour of the model isn't significantly modified
when adding the new unknown class.